Store sendto return in ssize_t data type
sendto returns ssize_t data type whereas this was getting stored in int
data type. Later rc is being logged with data type zd. This is causing
compilation error with ast2700 aarch64 compiler. Use more
suitable(ssize_t) data type to fix compilation error.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-22659

Tested:
Compilation is not failing after changing data type

+++ mctp-socket.c
@@ -146,7 +146,7 @@ mctp_requester_rc_t mctp_client_send(mctp_eid_t dest_eid, int mctp_fd,
 				     size_t req_msg_len)
 {
 	struct sockaddr_mctp addr = { 0 };
-	int rc = 0;
+	ssize_t rc = 0;
 
 	if (mctp_fd < 0) {
 		mctp_prerr("%s: Failed to create socket for mctp_fd %d: %s",
